+++ src/utils/input-validators/default-validator/default-validator.ts
@@ -1,0 +1,5 @@
+import { Validator } from '../validators.interface';
+
+export const defaultValidator: Validator<any> = {
+  validate: (_x: any) => true
+};