+++ src/utils/input-validators/validators.types.ts
@@ -1,0 +1,6 @@
+export enum ValidatorsTypes {
+  postCode = 'postcode',
+  licensePlate = 'license-plate',
+  moneyInput = 'money-input',
+  form = 'form'
+}